Retrieve a complete set of definitions

A function to retrieve all definitions and/or metadata for a given definition set. Caution: some datasets are large and may take time to retrieve.

Usage

catalog(type = NULL)

Arguments

type

character. The name of a definition set to retrieve. See Definitions section below for options.

Value

A data.frame containing an exhaustive list of the definitions and/or metadata retrievable for the given definition set. The variables in the data.frame will match those described in the documentation for that definition set.

Definitions

columns: Definitions of Macrostrat columns. See also def_columns().

drilling_sites: Definitions of eODP drilling cores. See also def_drilling_sites().

econs: Definitions of economic resources. See also def_econs().

environments: Definitions of environments. See also def_environments().

grain_sizes: Definitions of grain sizes. See also def_grain_sizes().

intervals: Definitions of temporal intervals. See also def_intervals().

lithologies: Definitions of lithologies. See also def_lithologies().

lithology_att: Definitions of lithology attributes. See also def_lithology_att().

measurements: Definitions of measurements. See also def_measurements().

minerals: Defintions of minerals. See also def_minerals().

plates: Definitions of tectonic plates. See also def_plates().

projects: Definitions of Macrostrat projects. See also def_projects().

references: Definitions of references. See also def_references().

sources: Definitions of geological maps. See also def_sources().

strat_name_concepts: Definitions of stratigraphic name concepts. See also def_strat_name_concepts().

strat_names: Definitions of stratigraphic names. See also def_strat_names().

structures: Definitions of geological structures. See also def_structures().

timescales: Definitions of timescales. See also def_timescales().

Developer(s)

Bethany Allen & Lewis A. Jones

Reviewer(s)

Lewis A. Jones

Examples

# \donttest{
  # Get mineral definitions
  ex1 <- catalog(type = "minerals")
  # Get grain size definitions
  ex2 <- catalog(type = "grain_sizes")
# }