Get lithology data from Macrostrat

This function gets lithology data from Macrostrat.

Usage

get_lithology(x = NULL, by = "name", available = FALSE)

Arguments

x

character. The name of a lithology available via Macrostrat.

by

character. The rank to search name by, should be either: "name", "group", "type", or "class".

available

logical. If set to TRUE, a vector of the available lithologies will be returned. If set to FALSE (default), the requested lithology will be checked against those available.

Value

A vector of available lithologies or a data.frame of the requested Macrostrat lithology containing at least the following columns:

lith_id

Lithology ID number.

name

Name of lithology.

type

Type of lithology.

group

Group of lithology.

class

Class of lithology.

color

Assigned lithology colour.

Details

This function returns a data.frame of the requested lithology and related data. Available Macrostrat lithologies can be requested via the available argument. Note: if this argument is set to TRUE, only a vector of available lithologies are returned. To return the desired lithology, this argument should be set to FALSE.

Examples

# Get lithology by name
get_lithology(by = "name", available = TRUE)
#>   [1] "siliciclastic"                 "gravel"                       
#>   [3] "sand"                          "silt"                         
#>   [5] "mud"                           "claystone"                    
#>   [7] "mudstone"                      "shale"                        
#>   [9] "siltstone"                     "sandstone"                    
#>  [11] "arkose"                        "greywacke"                    
#>  [13] "conglomerate"                  "breccia"                      
#>  [15] "diamictite"                    "mixed carbonate-siliciclastic"
#>  [17] "carbonate"                     "marl"                         
#>  [19] "lime mudstone"                 "wackestone"                   
#>  [21] "packstone"                     "grainstone"                   
#>  [23] "boundstone"                    "floatstone"                   
#>  [25] "rudstone"                      "bafflestone"                  
#>  [27] "bindstone"                     "framestone"                   
#>  [29] "limestone"                     "dolomite"                     
#>  [31] "siderite"                      "ankerite"                     
#>  [33] "evaporite"                     "halite"                       
#>  [35] "gypsum"                        "anhydrite"                    
#>  [37] "coal"                          "peat"                         
#>  [39] "lignite"                       "anthracite"                   
#>  [41] "tar"                           "chert"                        
#>  [43] "phosphorite"                   "ironstone"                    
#>  [45] "iron formation"                "volcaniclastic"               
#>  [47] "plutonic"                      "granite"                      
#>  [49] "monzonite"                     "granodiorite"                 
#>  [51] "syenite"                       "diorite"                      
#>  [53] "tonalite"                      "gabbro"                       
#>  [55] "foidolite"                     "peridotite"                   
#>  [57] "pyroxenite"                    "volcanic"                     
#>  [59] "rhyolite"                      "phonolite"                    
#>  [61] "trachyte"                      "rhyodacite"                   
#>  [63] "dacite"                        "andesite"                     
#>  [65] "basalt"                        "latite"                       
#>  [67] "felsite"                       "mafite"                       
#>  [69] "volcanic glass"                "ash"                          
#>  [71] "tuff"                          "welded tuff"                  
#>  [73] "metamorphic"                   "gneiss"                       
#>  [75] "migmatite"                     "mylonite"                     
#>  [77] "slate"                         "phyllite"                     
#>  [79] "schist"                        "quartzite"                    
#>  [81] "amphibolite"                   "greenstone"                   
#>  [83] "serpentinite"                  "hornfels"                     
#>  [85] "marble"                        "argillite"                    
#>  [87] "chalk"                         "clay"                         
#>  [89] "diabase"                       "igneous"                      
#>  [91] "ignimbrite"                    "till"                         
#>  [93] "greenschist"                   "quartz monzonite"             
#>  [95] "kimberlite"                    "anorthosite"                  
#>  [97] "carbonatite"                   "diatomite"                    
#>  [99] "calcareous ooze"               "troctolite"                   
#> [101] "bauxite"                       "lamprophyre"                  
#> [103] "norite"                        "granofel"                     
#> [105] "granulite"                     "sedimentary"                  
#> [107] "siliceous ooze"                "graywacke"                    
#> [109] "dolostone"                     "micrite"                      
#> [111] "greensand"                     "grit"                         
#> [113] "orthogneiss"                   "tephra"                       
#> [115] "paragneiss"                    "loess"                        
#> [117] "metasedimentary"               "coquina"                      
#> [119] "flint"                         "oolite"                       
#> [121] "trachyandesite"                "komatiite"                    
#> [123] "dunite"                        "charnockite"                  
#> [125] "aplite"                        "trondhjemite"                 
#> [127] "eclogite"                      "metapelite"                   
#> [129] "arenite"                       "pelite"                       
#> [131] "biomicrite"                    "oomicrite"                    
#> [133] "biosparite"                    "pelmicrite"                   
#> [135] "oosparite"                     "pelsparite"                   
#> [137] "intrasparite"                  "intramicrite"                 
#> [139] "monzogranite"                  "syenogranite"                 
#> [141] "novaculite"                    "travertine"                   
#> [143] "trona"                         "scoria"                       
#> [145] "pumice"                        "pegmatite"                    
#> [147] "dolerite"                      "encrinite"                    
#> [149] "alluvium"                      "colluvium"                    
#> [151] "regolith"                      "soil"                         
#> [153] "paleosol"                      "laterite"                     
#> [155] "eluvium"                       "mafic"                        
#> [157] "subarkose"                     "lherzolite"                   
#> [159] "harzburgite"                   "wehrlite"                     
#> [161] "websterite"                    "orthopyroxenite"              
#> [163] "clinopyroxenite"               "quartz arenite"               
#> [165] "metabasalt"                    "metagabbro"                   
#> [167] "cataclasite"                   "fault gouge"                  
#> [169] "fault breccia"                 "pseudotachylite"              
#> [171] "blueschist"                    "skarn"                        
#> [173] "metavolcanic"                  "obsidian"                     
#> [175] "gyttja"                        "drift"                        
#> [177] "tillite"                       "metaconglomerate"             
#> [179] "diatexite"                     "ankaramite"                   
#> [181] "wacke"                         "agglomerate"                  
#> [183] "leucogranite"                  "metaigneous"                  
#> [185] "tufa"                          "metarhyolite"                 
#> [187] "metagraywacke"                 "bentonite"                    
#> [189] "metasiltstone"                 "basanite"                     
#> [191] "tuffite"                       "litharenite"                  
#> [193] "sublitharenite"                "radiolarite"                  
#> [195] "calcarenite"                   "calcilutite"                  
#> [197] "metabasite"                    "diamicton"                    
#> [199] "porcellanite"                  "hyaloclastite"                
#> [201] "hawaiite"                      "mugearite"                    
#> [203] "granophyre"                    "comendite"                    
#> [205] "tephrite"                      "spilite"                      
#> [207] "benmoreite"                   
# Get lithology by type
get_lithology(by = "type", available = TRUE)
#>  [1] "siliciclastic"   "carbonate"       "evaporite"       "organic"        
#>  [5] "chemical"        "plutonic"        "volcanic"        "metamorphic"    
#>  [9] "metasedimentary" "metavolcanic"    "igneous"         "sedimentary"    
#> [13] "metaigneous"     "regolith"        "cataclastic"    
# Get lithology by group
get_lithology(by = "group", available = TRUE)
#> [1] "unconsolidated" "mudrocks"       "sandstones"     "conglomerates" 
#> [5] "Dunham"         "felsic"         "mafic"          "ultramafic"    
#> [9] "Folk"          
# Get lithology by class
get_lithology(by = "class", available = TRUE)
#> [1] "sedimentary" "igneous"     "metamorphic"
# Get all sedimentary lithologies
get_lithology(x = "sedimentary", by = "class", available = FALSE)
#>    lith_id                          name          type          group
#> 1        1                 siliciclastic siliciclastic           <NA>
#> 2        2                        gravel siliciclastic unconsolidated
#> 3        3                          sand siliciclastic unconsolidated
#> 4        4                          silt siliciclastic unconsolidated
#> 5        5                           mud siliciclastic unconsolidated
#> 6        6                     claystone siliciclastic       mudrocks
#> 7        7                      mudstone siliciclastic       mudrocks
#> 8        8                         shale siliciclastic       mudrocks
#> 9        9                     siltstone siliciclastic       mudrocks
#> 10      10                     sandstone siliciclastic     sandstones
#> 11      11                        arkose siliciclastic     sandstones
#> 12      12                     greywacke siliciclastic     sandstones
#> 13      14                  conglomerate siliciclastic  conglomerates
#> 14      15                       breccia siliciclastic  conglomerates
#> 15      16                    diamictite siliciclastic  conglomerates
#> 16      17 mixed carbonate-siliciclastic     carbonate           <NA>
#> 17      18                     carbonate     carbonate           <NA>
#> 18      19                          marl     carbonate           <NA>
#> 19      20                 lime mudstone     carbonate         Dunham
#> 20      21                    wackestone     carbonate         Dunham
#> 21      22                     packstone     carbonate         Dunham
#> 22      23                    grainstone     carbonate         Dunham
#> 23      24                    boundstone     carbonate         Dunham
#> 24      25                    floatstone     carbonate         Dunham
#> 25      26                      rudstone     carbonate         Dunham
#> 26      27                   bafflestone     carbonate         Dunham
#> 27      28                     bindstone     carbonate         Dunham
#> 28      29                    framestone     carbonate         Dunham
#> 29      30                     limestone     carbonate           <NA>
#> 30      31                      dolomite     carbonate           <NA>
#> 31      32                      siderite     carbonate           <NA>
#> 32      33                      ankerite     carbonate           <NA>
#> 33      34                     evaporite     evaporite           <NA>
#> 34      35                        halite     evaporite           <NA>
#> 35      36                        gypsum     evaporite           <NA>
#> 36      37                     anhydrite     evaporite           <NA>
#> 37      38                          coal       organic           <NA>
#> 38      39                          peat       organic           <NA>
#> 39      40                       lignite       organic           <NA>
#> 40      43                    anthracite       organic           <NA>
#> 41      44                           tar       organic           <NA>
#> 42      45                         chert      chemical           <NA>
#> 43      46                   phosphorite      chemical           <NA>
#> 44      47                     ironstone      chemical           <NA>
#> 45      48                iron formation      chemical           <NA>
#> 46      49                volcaniclastic siliciclastic           <NA>
#> 47      91                     argillite siliciclastic       mudrocks
#> 48      92                         chalk     carbonate           <NA>
#> 49      93                          clay siliciclastic unconsolidated
#> 50      97                          till siliciclastic  conglomerates
#> 51     103                     diatomite      chemical           <NA>
#> 52     104               calcareous ooze     carbonate           <NA>
#> 53     106                       bauxite      chemical           <NA>
#> 54     111                   sedimentary   sedimentary           <NA>
#> 55     112                siliceous ooze      chemical           <NA>
#> 56     113                     graywacke siliciclastic     sandstones
#> 57     114                     dolostone     carbonate           <NA>
#> 58     115                       micrite     carbonate           Folk
#> 59     116                     greensand siliciclastic     sandstones
#> 60     117                          grit siliciclastic     sandstones
#> 61     121                         loess siliciclastic           <NA>
#> 62     123                       coquina     carbonate           <NA>
#> 63     124                         flint      chemical           <NA>
#> 64     125                        oolite     carbonate           <NA>
#> 65     134                       arenite siliciclastic     sandstones
#> 66     135                        pelite siliciclastic       mudrocks
#> 67     136                    biomicrite     carbonate           Folk
#> 68     137                     oomicrite     carbonate           Folk
#> 69     138                    biosparite     carbonate           Folk
#> 70     139                    pelmicrite     carbonate           Folk
#> 71     140                     oosparite     carbonate           Folk
#> 72     141                    pelsparite     carbonate           Folk
#> 73     142                  intrasparite     carbonate           Folk
#> 74     143                  intramicrite     carbonate           Folk
#> 75     146                    novaculite      chemical           <NA>
#> 76     147                    travertine     carbonate           <NA>
#> 77     148                         trona     evaporite           <NA>
#> 78     153                     encrinite     carbonate           <NA>
#> 79     154                      alluvium siliciclastic unconsolidated
#> 80     155                     colluvium siliciclastic unconsolidated
#> 81     156                      regolith      regolith           <NA>
#> 82     157                          soil      regolith           <NA>
#> 83     158                      paleosol      regolith           <NA>
#> 84     159                      laterite      regolith           <NA>
#> 85     160                       eluvium      regolith           <NA>
#> 86     162                     subarkose siliciclastic     sandstones
#> 87     169                quartz arenite siliciclastic     sandstones
#> 88     180                        gyttja       organic unconsolidated
#> 89     181                         drift siliciclastic unconsolidated
#> 90     182                       tillite siliciclastic  conglomerates
#> 91     187                         wacke siliciclastic     sandstones
#> 92     191                          tufa     carbonate           <NA>
#> 93     198                   litharenite siliciclastic     sandstones
#> 94     199                sublitharenite siliciclastic     sandstones
#> 95     200                   radiolarite      chemical           <NA>
#> 96     201                   calcarenite     carbonate           <NA>
#> 97     202                   calcilutite     carbonate           <NA>
#> 98     204                     diamicton siliciclastic unconsolidated
#> 99     205                  porcellanite      chemical           <NA>
#>          class   color fill t_units
#> 1  sedimentary #FFF400  670     346
#> 2  sedimentary #FFAB00  601    1188
#> 3  sedimentary #FFCA00  607   11704
#> 4  sedimentary #919AA3  619    8596
#> 5  sedimentary #A7ACB0  641    8274
#> 6  sedimentary #B2B5B7  620    4742
#> 7  sedimentary #9CA3AA  620    4875
#> 8  sedimentary #7B8896  620   11246
#> 9  sedimentary #86919D  616    6708
#> 10 sedimentary #FFD500  607   12874
#> 11 sedimentary #FFEA00  654     407
#> 12 sedimentary #FFDF00  654     486
#> 13 sedimentary #FFC000  602    3714
#> 14 sedimentary #FFB500  603     719
#> 15 sedimentary #FF9600  682     209
#> 16 sedimentary #0000FF  674     230
#> 17 sedimentary #0508FD  641     742
#> 18 sedimentary #0B11FC  623     654
#> 19 sedimentary #111AFB  627      66
#> 20 sedimentary #1723FA  627    1399
#> 21 sedimentary #1D2BF9  628    1814
#> 22 sedimentary #2334F8  629     930
#> 23 sedimentary #293DF7  627      37
#> 24 sedimentary #2F46F6  627     267
#> 25 sedimentary #344EF5    0     297
#> 26 sedimentary #3A57F4    0       0
#> 27 sedimentary #4060F3    0       5
#> 28 sedimentary #4669F2    0      78
#> 29 sedimentary #4C71F1  627    9590
#> 30 sedimentary #527AF0  642    4155
#> 31 sedimentary #5883EF    0       8
#> 32 sedimentary #5E8CEE    0       0
#> 33 sedimentary #FFB6C1  667     101
#> 34 sedimentary #F9AFBA  668     213
#> 35 sedimentary #F3A8B3  667     214
#> 36 sedimentary #EEA2AD  667     569
#> 37 sedimentary #000000  658     670
#> 38 sedimentary #090F0F  657      33
#> 39 sedimentary #121F1F  657      94
#> 40 sedimentary #1C2F2F  658       2
#> 41 sedimentary #253F3F    0       0
#> 42 sedimentary #CD5C5C  649    2101
#> 43 sedimentary #C96566    0      29
#> 44 sedimentary #C66F70    0      45
#> 45 sedimentary #C37879    0      73
#> 46 sedimentary #FFFF00  714    3180
#> 47 sedimentary #708090    0     554
#> 48 sedimentary #6495ED  626    4428
#> 49 sedimentary #BEBEBE    0   25047
#> 50 sedimentary #FFA000  682     151
#> 51 sedimentary #BF8283  653     112
#> 52 sedimentary #BC8B8E  626   14271
#> 53 sedimentary #B99598    0      13
#> 54 sedimentary #FF8C00    0      71
#> 55 sedimentary #CCEDCC    0    3048
#> 56 sedimentary #FFDF00  654      72
#> 57 sedimentary #527AF0    0     332
#> 58 sedimentary #111AFB    0      21
#> 59 sedimentary #E6C000    0       0
#> 60 sedimentary #FFD500    0       4
#> 61 sedimentary #919AA3    0       0
#> 62 sedimentary #2334F8    0     262
#> 63 sedimentary #CD5C5C    0       0
#> 64 sedimentary #2334F8    0       7
#> 65 sedimentary #FFD500    0      71
#> 66 sedimentary #9CA3AA    0       7
#> 67 sedimentary #111AFB    0       0
#> 68 sedimentary #111AFB    0       0
#> 69 sedimentary #2334F8    0       0
#> 70 sedimentary #111AFB    0       0
#> 71 sedimentary #2334F8    0       0
#> 72 sedimentary #2334F8    0       0
#> 73 sedimentary #2334F8    0       0
#> 74 sedimentary #111AFB    0       0
#> 75 sedimentary #CD5C5C    0       0
#> 76 sedimentary #1D2BF9    0       2
#> 77 sedimentary #EEA2AD    0       1
#> 78 sedimentary #2334F8  629       0
#> 79 sedimentary #FFCA00  101      34
#> 80 sedimentary #FFCA00  102       0
#> 81 sedimentary #FFFF80    0       0
#> 82 sedimentary #FFE6E6    0       6
#> 83 sedimentary #FFE6E6    0       1
#> 84 sedimentary #FFE6E6    0       5
#> 85 sedimentary #FFE6E6    0       0
#> 86 sedimentary #FFEA00  654      47
#> 87 sedimentary #FFD500    0     525
#> 88 sedimentary #0F0301    0       0
#> 89 sedimentary #FFAB00  601       0
#> 90 sedimentary #E08D02  682       2
#> 91 sedimentary #FFDF00  654      63
#> 92 sedimentary #1D2BF9    0       0
#> 93 sedimentary #FFDF00  654      78
#> 94 sedimentary #FFDF00  654      35
#> 95 sedimentary #BF8283  653      25
#> 96 sedimentary #2334F8  629      23
#> 97 sedimentary #111AFB  627      14
#> 98 sedimentary #FF9600  682     555
#> 99 sedimentary #CD5C5C  649     230
# Get specific lithology
get_lithology(x = "limestone", by = "name", available = FALSE)
#>   lith_id      name      type group       class   color fill t_units
#> 1      30 limestone carbonate    NA sedimentary #4C71F1  627    9590