FUNCTION_TITLE

FUNCTION_DESCRIPTION

Usage

get_sections(
  col_id,
  unit_id,
  section_id,
  interval_name,
  int_id,
  age,
  age_top,
  age_bottom,
  lith,
  lith_id,
  lith_group,
  lith_type,
  lith_class,
  lith_att,
  lith_att_id,
  lith_att_type,
  environ,
  environ_id,
  environ_type,
  environ_class,
  econ,
  econ_id,
  econ_type,
  econ_class,
  cltn_id,
  strat_name,
  strat_name_id,
  lat,
  lng,
  adjacents,
  project_id,
  response = NULL
)

Arguments

col_id

integer. The unique identification number(s) of the columns(s) to return.

unit_id

integer. The unique identification number(s) of the unit(s) to filter the results by.

section_id

integer. The unique identification number(s) of the section(s) to filter the results by.

interval_name

character. The name of a chronostratigraphic time interval to filter the results by.

int_id

integer. The unique identification number(s) of a chronostratigraphic time interval to filter the results by.

age

numeric. Filter sections to those that overlap this numerical age in millions of years before present.

age_top

numeric. Filter sections to those that overlap with the age range between this and age_bottom. Should be in millions of years before present and less than age_bottom.

age_bottom

numeric. Filter sections to those that overlap with the age range between this and age_top. Should be in millions of years before present and greater than age_top.

lith

character. Filter sections to those containing a named lithology (e.g., shale, sandstone).

lith_id

integer. Filter sections to those containing a lithology identified by its unique identification number.

lith_group

character. Filter sections to those containing a named lithology group (e.g., sandstones, mudrocks, unconsolidated)

lith_type

character. Filter sections to those containing a named lithology type (e.g., carbonate, siliciclastic).

lith_class

character. Filter sections to those containing a named general lithology (sedimentary, igneous, metamorphic).

lith_att

character. Filter sections to those containing a named lithology attribute (e.g. fine, olivine, poorly washed).

lith_att_id

integer. Filter sections to those containing a lithology attribute identified by its unique identification number.

lith_att_type

character. Filter sections to those containing a named category of lithology attribute (e.g. grains, lithology, bedform).

environ

character. Filter sections to those containing a named environment.

environ_id

integer. Filter sections to those containing an environment identified by its unique identification number.

environ_type

character. Filter sections to those containing a named group of environments.

environ_class

character. Filter sections to those containing a named general environment.

econ

character. Filter sections to those containing a named economic attribute.

econ_id

integer. Filter sections to those containing an economic attribute identified by its unique identification number.

econ_type

character. Filter sections to those containing a named economic attribute type.

econ_class

character. Filter sections to those containing a named economic attribute class.

cltn_id

integer. Filter sections to those containing one or more Paleobiology Database collections identified by their unique identification numbers.

strat_name

character. Filter sections to those containing a unit that fuzzy matches a stratigraphic name.

strat_name_id

integer. Filter sections by a single or comma-separated list of stratigraphic IDs.

lat

numeric. Return the section at a particular decimal degree latitude (WGS84). Must also specify lng.

lng

numeric. Return the section at a particular decimal degree longitude (WGS84). Must also specify lat.

adjacents

logical. If lat/lng or col_id is specified, should all sections that touch the specified column be returned?

project_id

integer. Filter sections to those contained within a Macrostrat project identified by its unique identification number.

response

character. How much metadata should be returned? Options (in increasing amount of metadata) are "short" and "long"., Default: NULL

Value

OUTPUT_DESCRIPTION

Details

DETAILS

Author

AUTHOR AUTHOR_2

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}