Get Macrostrat project statistics from Macrostrat

This function gets specific statistics on Macrostrat projects from Macrostrat.

Usage

get_stats(x = NULL, available = FALSE)

Arguments

x

character. The name of a project available via Macrostrat.

available

logical. If set to TRUE, a vector of the available projects will be returned. If set to FALSE (default), the requested project will be checked against those available.

Value

A vector of available projects or a data.frame of the requested Macrostrat project statistics containing:

project_id

The project ID number.

project

The name of the project.

columns

The number of geological columns associated with the project.

packages

The number of geological packages associated with the project.

units

The number of geological units associated with the project.

pbdb_collections

The number of Paleobiology Database collections associated with the project.

measurements

The number of measurements associated with the project.

t_polys

The number of polygons(?) associated with the project.

Details

This function returns a data.frame of the requested project stats. Available Macrostrat project statistics can be requested via the available argument. Note: if this argument is set to TRUE, only a vector of available project names are returned. To return the desired project statistics, this argument should be set to FALSE.

Examples

# Get available timescales
available <- get_stats(available = TRUE)
#> Error: Please install readr package
# Get all stats
all <- get_stats()
#> Error: Please install readr package
# Get project specific
nz <- get_stats(x = "New Zealand")
#> Error: Please install readr package