Get timescale data from Macrostrat

This function gets timescale data from Macrostrat.

get_timescale(x = "international ages", available = FALSE)

Arguments

x

character. The name of a timescale available via Macrostrat.

available

logical. If set to TRUE, a vector of the available timescales will be returned. If set to FALSE (default), the requested scale will be checked against those available.

Value

A vector of available timescales or a data.frame of the requested Macrostrat timescale containing:

name

Interval name.

abbrev

Interval name abbreviation.

t_age

Interval top age, i.e. minimum age.

b_age

Interval bottom age, i.e. maximum age.

color

Assigned interval colour.

Details

This function returns a data.frame of the requested timescale. Available Macrostrat timescales can be requested via the available argument. Note: if this argument is set to TRUE, only a vector of available scales are returned. To return the desired timescale, this argument should be set to FALSE.

Examples

# Get available timescales
available <- get_timescale(available = TRUE)
#> Error: Please install readr package
# Get timescale data
ages <- get_timescale(x = "international ages")
#> Error: Please install readr package
periods <- get_timescale(x = "international periods")
#> Error: Please install readr package