Retrieve Macrostrat hosted interval data

This function fetches Macrostrat interval data based on user-defined arguments via the Macrostrat API.

Usage

macrostrat_intervals(
  timescale = NULL,
  name = NULL,
  age = NULL,
  t_age = NULL,
  b_age = NULL,
  rule = NULL,
  all = TRUE,
  true_colors = NULL,
  int_id = NULL,
  timescale_id = NULL
)

Arguments

timescale

character. The name of the desired timescale. Available timescales can be retrieved via macrostrat_timescales.

name

character. The name of the desired interval. If set to NULL (default), all interval data for the respective timescale is returned.

age

numeric. Age in millions of years before present. All intervals overlapping with this age will be returned.

t_age

numeric. The minimum age bound interval data should be returned for in millions of years before present. If set to NULL (default), all interval data for the respective timescale is returned.

b_age

numeric. The maximum age bound interval data should be returned for in millions of years before present. If set to NULL (default), all interval data for the respective timescale is returned.

rule

character. How should intervals be returned for the given t_age and b_age? Use "contains" to return all intervals that fall entirely within t_age and b_age. Use "exact" to return all intervals within boundaries equal to t_age and b_age. Use "loose" to return all intervals touching the range of t_age and b_age. If t_age and b_age are not provided, this argument is ignored.

all

logical. Should all available interval definitions be returned? Defaults to TRUE.

true_colors

logical. Should the original international time scale colors be returned? Defaults to TRUE.

int_id

integer. The identification number of the desired interval.

timescale_id

integer. The identification number of the desired timescale. Available timescale identification numbers can be retrieved via macrostrat_timescales.

Value

A data.frame of time scales hosted by Macrostrat containing:

int_id: The unique identification number of the interval.

name: The name of the interval.

abbrev: If available, the interval abbreviation.

t_age: The minimum age of the interval.

b_age: The maximum age of the interval.

int_type: The type (or rank) of interval (e.g. age, period).

timescales: The time scale the interval belongs to.

color: International time scale colors.

Details

DETAILS

Author

Lewis A. Jones

Examples

# Get international ages
ages <- macrostrat_intervals(timescale = "international ages")
# Get international periods
periods <- macrostrat_intervals(timescale = "international periods")
# Get NALMAs
NALMAs <- macrostrat_intervals(timescale = "North American Land Mammal Ages")
# Get specific interval by name
Hettangian <- macrostrat_intervals(name = "Hettangian")
# Get specific interval by ID
Chattian <- macrostrat_intervals(int_id = 20L)
# Get international age by age
Gzhelian <- macrostrat_intervals(timescale = "international ages", age = 300)
# Get international age by age range
intervals <- macrostrat_intervals(timescale = "international ages", t_age = 0, b_age = 100)