Retrieve Macrostrat hosted time scales

This function fetches all available time scales from Macrostrat via the Macrostrat API.

Usage

macrostrat_timescales()

Value

A data.frame of time scales hosted by Macrostrat containing:

timescale_id: The time scale identification number.

timescale: The name of the time scale.

n_intervals: The number of intervals within the time scale.

max_age: The maximum age of the intervals in the time scale.

min_age: The minimum age of the intervals in the time scale.

ref_id: The identification number of the reference the time scale is based on.

Author

Lewis A. Jones

Examples

# Get available time scales
macrostrat_timescales()
#> Warning: Coercing LHS to a list
#> # A tibble: 37 × 6
#>    timescale_id timescale                     n_intervals max_age min_age ref_id
#>           <dbl> <chr>                               <dbl>   <dbl>   <dbl>  <dbl>
#>  1            1 international ages                    102    539.  0          13
#>  2            2 international epochs                   34    539.  0          13
#>  3            3 international periods                  22   2500   0          13
#>  4            5 calcareous nannoplankton zon…          88    201.  0           0
#>  5            6 New Zealand ages                       89    512.  0           0
#>  6            7 custom COSUNA                          84    539.  0.0117      0
#>  7            9 North American land mammal a…          19     66   0.0114      0
#>  8           11 international intervals               170   4000   0          13
#>  9           12 COSUNA                                 62    539.  0           0
#> 10           13 international eras                     10   4000   0          13
#> # ℹ 27 more rows