rmacrostrat: Fetch Geologic Data from the 'Macrostrat' Platform

Work with the 'Macrostrat' (https://macrostrat.org/) Web Service (v.2, https://macrostrat.org/api/v2) to fetch geological data relevant to the spatial and temporal distribution of sedimentary, igneous, and metamorphic rocks as well as data extracted from them.

See also

Author

Maintainer: Lewis A. Jones LewisA.Jones@outlook.com (ORCID) [copyright holder]

Authors:

William Gearty willgearty@gmail.com (ORCID)

Christopher D. Dean christopherdaviddean@gmail.com (ORCID)

Bethany Allen bethany.allen@bsse.ethz.ch (ORCID)