Retrieve column age models

Obtain information about the age models for columns. This largely consists of the temporal boundaries applied to the column's constituent units.

Usage

get_age_model(column_id = NULL, section_id = NULL)

Arguments

column_id

integer. Filter age models to those associated with column(s) as specified by their unique identification number(s).

section_id

integer. Filter age models to those associated with section(s) as specified by their unique identification number(s).

Value

A dataframe containing, for each boundary between units for each retrieved age model:

boundary_id

The unique identifier of the boundary between two units

col_id

The unique identifier of the Macrostrat column

section_id

The unique identifier of the Macrostrat section

interval_id

The unique identifier of the time interval which the boundary lies within

interval_name

The name of the time interval which the boundary lies within

age_bottom

The numerical lower limit on the age of the boundary

age_top

The numerical upper limit on the age of the boundary

rel_position

The relative position of the boundary

model_age

The age of the boundary, as correlated with other ages in the column

boundary_status

The method used to determine the age of the boundary

boundary_type

The type of boundary

boundary_position

The position of the boundary

unit_below

The Macrostrat unit beneath the given boundary

unit_above

The Macrostrat unit above the given boundary

ref_id

The unique identifier of the reference

Details

The Macrostrat age models are used to determine the ages of the boundaries between units in an internally consistent manner, allowing each boundary to be dated with respect to all age information available for the column.

Author

Bethany Allen

Examples

if (FALSE) {
if (interactive()) {
  # age_model <- get_age_model(column_id = c(503, 504))
}
}