Retrieve Macrostrat hosted time scales

This function fetches all available time scales from Macrostrat via the Macrostrat API.

Usage

macrostrat_timescales()

Value

A data.frame of time scales hosted by Macrostrat containing:

timescale_id: The time scale identification number.

timescale: The name of the time scale.

n_intervals: The number of intervals within the time scale.

max_age: The maximum age of the intervals in the time scale.

min_age: The minimum age of the intervals in the time scale.

ref_id: The identification number of the reference the time scale is based on.

Author

Lewis A. Jones

Examples

# Get available time scales
macrostrat_timescales()
#> Warning: Coercing LHS to a list
#>    timescale_id                                           timescale n_intervals
#> 1             1                                  international ages         102
#> 2             2                                international epochs          34
#> 3             3                               international periods          22
#> 4             5                      calcareous nannoplankton zones          88
#> 5             6                                    New Zealand ages          89
#> 6             7                                       custom COSUNA          84
#> 7             9                     North American land mammal ages          19
#> 8            11                             international intervals         170
#> 9            12                                              COSUNA          62
#> 10           13                                  international eras          10
#> 11           14                                  international eons           3
#> 12           15                      Trilobite Zonation - Laurentia          25
#> 13           16                                   Conodont Zonation         198
#> 14           17                             North American Regional          95
#> 15           18                          Ammonite Zonation - Boreal         154
#> 16           19                Ammonite Zonation - Western Interior          70
#> 17           20           international intervals covering all time          16
#> 18           21                              Scotese Reconstruction          17
#> 19           22                          Geomagnetic Polarity Chron         101
#> 20           23                       Geomagnetic Polarity Subchron         372
#> 21           24             Planktic foraminiferal Primary Biozones          51
#> 22           25           Planktic foraminiferal Secondary Biozones          39
#> 23           26                       Planktic foraminiferal datums           2
#> 24           27                                     Martian Periods           4
#> 25           28                                      Martian Epochs           8
#> 26           29           Cretaceous Planktic foraminifer zonations          31
#> 27           30                   Low latitude radiolarian zonation          34
#> 28           31          Neogene North Pacific Diatom Biochronology          11
#> 29           32 Neogene North Pacific Diatom Biochronology Subzones          22
#> 30           33                                   Siberian Regional          22
#> 31           34                                 Australian Regional          21
#> 32           35                             Western Europe Regional          21
#> 33           36                    Russian Platform Regional Stages          51
#> 34           37                            Russian Precambrian Eras           8
#> 35           38                            Russian Precambrian Eons           4
#> 36           39                                      Russian Epochs          36
#> 37           40                                      Russian Stages          97
#>    max_age  min_age ref_id
#> 1   538.80   0.0000     13
#> 2   538.80   0.0000     13
#> 3  2500.00   0.0000     13
#> 4   201.40   0.0000      0
#> 5   511.50   0.0000      0
#> 6   538.80   0.0117      0
#> 7    66.00   0.0114      0
#> 8  4000.00   0.0000     13
#> 9   538.80   0.0000      0
#> 10 4000.00   0.0000     13
#> 11 4000.00   0.0000     13
#> 12  524.00 484.4000     13
#> 13  485.40 201.4000     13
#> 14  850.00   0.0114     13
#> 15  252.40  97.5000     13
#> 16  100.25  68.6900      0
#> 17 4000.00   0.0000     13
#> 18  113.00   0.0000     28
#> 19  170.76   0.0000    180
#> 20  170.76   0.0000    180
#> 21   65.00   0.0000    190
#> 22   40.80   0.0000    190
#> 23   59.20  56.3000    190
#> 24 4600.00   0.0000    204
#> 25 4180.00   0.0000    204
#> 26  141.60  65.0000    219
#> 27   60.20   0.0000    220
#> 28   24.00   0.0000    221
#> 29   24.00   0.0000    221
#> 30  600.00 297.0000    222
#> 31  511.50 433.0000    222
#> 32  362.00 274.3700    222
#> 33  360.00 251.9020    222
#> 34 3200.00 600.0000    224
#> 35 4000.00 600.0000    224
#> 36  600.00   0.0000    224
#> 37  538.80   0.0117    224